Update: UserRepository [카드 조회 where 학교 추가 및 동적 쿼리 단위 테스트], Fix: UserRepository [카드 조회 left join 으로 인해 팔로우 수 1명과 0명이 구분이 안되는 부분 수정]

+++ src/main/java/okky/team/chawchaw/user/dto/UserCardDto.java
@@ -18,6 +18,6 @@ public class UserCardDto {
     String repHopeLanguage;
     LocalDateTime days;
     Long views;
-    Long follows;
+    Integer follows;
 
 }
